feat: create error handling
