+++ web/src/reducers/play.ts
@@ -2,6 +2,8 @@ const SET_LAST_WORD = "SET_LAST_WORD" as const;
 const SET_INPUT_WORD = "SET_INPUT_WORD" as const;
 const VERIFY_JAPANESE_WORD = "VERIFY_JAPANESE_WORD" as const;
 const SET_ERROR = "SET_ERROR" as const;
+const CHECK_WORD_ERROR = "CHECK_WORD_ERROR" as const;
+const SET_WORD_ERROR_MESSAGE = "SET_WORD_ERROR_MESSAGE" as const;
 
 const setLastWord = (word: string) => {
   return { type: SET_LAST_WORD, word: word };
@@ -19,24 +21,39 @@ const setError = (error?: Error) => {
   return { type: SET_ERROR, error: error };
 };
 
+const checkWordError = (hasWordError: boolean) => {
+  return { type: CHECK_WORD_ERROR, hasWordError: hasWordError };
+};
+
+const setWordErrorMessage = (wordErrorMessage: string) => {
+  return { type: SET_WORD_ERROR_MESSAGE, wordErrorMessage: wordErrorMessage };
+};
+
+
 export const actions = {
   setLastWord,
   setInputWord,
   verifyJapaneseWord,
   setError,
+  checkWordError,
+  setWordErrorMessage
 };
 
 export type Actions =
   | ReturnType<typeof setLastWord>
   | ReturnType<typeof setInputWord>
   | ReturnType<typeof verifyJapaneseWord>
-  | ReturnType<typeof setError>;
+  | ReturnType<typeof setError>
+  | ReturnType<typeof checkWordError>
+  | ReturnType<typeof setWordErrorMessage>;
 
 export type State = {
   lastWord: string;
   inputWord: string;
   isValidJapanese: boolean;
   error?: Error;
+  hasWordError: boolean;
+  wordErrorMessage: string;
 };
 
 export const initialState: State = {
@@ -44,6 +61,8 @@ export const initialState: State = {
   inputWord: "",
   isValidJapanese: false,
   error: undefined,
+  hasWordError: false,
+  wordErrorMessage: "",
 };
 
 export const reducer = (state: State, action: Actions): State => {
@@ -67,6 +86,16 @@ export const reducer = (state: State, action: Actions): State => {
       return {
         ...state,
         error: action.error,
+      };
+    case CHECK_WORD_ERROR:
+      return {
+        ...state,
+        hasWordError: action.hasWordError,
+      };
+    case SET_WORD_ERROR_MESSAGE:
+      return {
+        ...state,
+        wordErrorMessage: action.wordErrorMessage,
       };
     default:
       return state;
